Half-Functioning Heart Rate calculator.
The display works well, but to determine a heartbeat from the data and then calculate the heart rate from it may be too computationally heavy for practical purposes. Right now there is a system to record the average of the previous 8 values displayed, but the heartBeat trigger still depends on the voltage crossing over a certain threshold. Both systems have their merits and their flaw. The threshold is not completely consistent, and the moving average slows down the process of recording by a noticeable amount.  